Refactored to remove circular dependencies, and added post-invoke logic
